requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3

# Pydantic
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# OTP delivery
fastapi-mail==1.4.1

# Firebase Cloud Messaging
google-auth==2.29.0

# File uploads
python-multipart==0.0.12
aiofiles==24.1.0

# HTTP client
httpx==0.27.2

# Dev / testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
psycopg2-binary==2.9.11